Added arrow assets
At this point, I realized that to move on, I needed to add the arrow
pictures to move on and add some functionality. The arrows will
act as buttons so that you can navigate through your gallery.

It's unformatted right now, I'll style the arrows so they're not so
large in the next commit.