Added ability to change main picture
If you're just figuring out how to do things - this is a great way to
do it: in small steps. In this step, I needed to test if I could make a
button change the main picture. Turns out I could do it! Here's how I
did it:

First, I relalized that I needed to put identifiers in my HTML so
that my javascript knows what is happening. I put id="back" and
id="forward" on my arrow kews so that my javascript can know that
those are the forward and back keys. I also put id="main-picture"
in the img tag that contains the main picture so my javascript
code knows what is happening.

You don't have to do this first; you can try to write your javascript
first and then realize you need to add these later on.

After that, I write a function that would switch the 'src' property of
my img tag to a new url. It's a little dense, what's happening is I tell
the comptuer to look for the "main-picture" identifier. Whenever the
changePicture() function is called then, the computer switches the src
property in main-picture to the new one I defined before.

Finally, I needed to make this happen when you clicked the forward
button. This is handled by the event handler routine that I call -
it looks for the identifier called "forward" and binds whatever you want
to it. In this case, I put the changePicture() function in it so that it
gets called everytime you click the element that contains the identifier
"forward".