Added ability to change titles and links
While navigating through our gallery, we want users to be able to
read a different title for each picture as well as have a link at the
bottom so that they can navigate to the picture directly.

The process is similar to what we did in the last commit. Here, I
added identifiers to the picture title and the picture link so that
while I'm writing my javascript I can modify the tags.

The only change I needed to make in the javascript was to add
functions to modify three things:

  * The page title
  * The page link
  * The page link text