Added ability to go forward through multiple pics
We're building a gallery right? That means we want people to
be able to cycle through multiple pictures. Here, I provided some
basic functionality - we can cycle forward through multiple pictures.

This uses a couple of things that are a little more complicated;
arrays and objects.

I added an array where we can store multiple pictures and their titles.
In each element of the array, we have a 'url' and a 'title'. They're
pretty much self explainatory - they're the URL and the TITLE of the
pictures you want to show.

Objects are denoted by braces "{}" and arrays are denoted by square
brackets "[]". Within them, each element is separated by a comma.

Below, you can see that when the changePicture function is called,
there are now several functions which grab the pictures using an index;
If you remember correctly, we need an index every time we access an
element in an array. It was simple to substitute this in for the
code that was there before!

Also note - we update the index at the bottom so that the next time
you hit the forward button, the index updates and the browser shows the
next picture.