Added min and max heighs for main picture
We have most of our functionality down now - it's now time to refine and
make it pretty.

If you flipped through the pictures in the previous commit, you could
tell that it got a little annoying when you had huge pictures. I fixed
this by adding a min-height and max-height property to the main picture
in the CSS file.

These two lines make it so that the height for all pictures is constant.