Added checks for gallery bounds
There was a bug earlier on that allowed you to
keep hitting back even though you were at the beginning of the
gallery. Here, we include a check for each button click to see
if we're at the end of the start of the gallery.

If we're at the end or the start, we alert the user and don't increment
the index.